feat(card-component): Add Card Component
This component was created to manage all of the card-related processes. Card name, description, and artwork can be accessed through the component and use anywhere. The component is storing cards in two different lists: permanent card lists and temporal card lists. All of the info can be accessed via giving an index number to this component's methods.

Co-Authored-By: Eren Kumru <44412775+ErenKumru@users.noreply.github.com>